Implement average values as doubles (#110)
Having average value calculations as unsigned integers was producing harsh
rounding errors for small valued quantities (like nprocs, nthreads). These are
now implemented as double precision values and get serialised as
numbers in JSON with decimal fractions.

There is a slight drawback that the decimal part is designed in nlohmann::json
to avoid precision loss, so sometimes a very high number of DPs are printed.
However, there doesn't seem to be any way to adjust this without manually
massaging the actual numerical values (there are various issues in
https://github.com/nlohmann/json where this is discussed).

+++ package/src/countmon.h
@@ -1,4 +1,4 @@
-// Copyright (C) CERN, 2018
+// Copyright (C) CERN, 2020
 //
 // Process and thread number monitoring class
 //
@@ -21,7 +21,7 @@ class countmon final : public Imonitor {
   // Container for total stats
   std::map<std::string, unsigned long long> count_stats;
   std::map<std::string, unsigned long long> count_peak_stats;
-  std::map<std::string, unsigned long long> count_average_stats;
+  std::map<std::string, double> count_average_stats;
   std::map<std::string, unsigned long long> count_total_stats;
 
   // Counter for number of iterations
@@ -35,7 +35,7 @@ class countmon final : public Imonitor {
   // These are the stat getter methods which retrieve current statistics
   std::map<std::string, unsigned long long> const get_text_stats();
   std::map<std::string, unsigned long long> const get_json_total_stats();
-  std::map<std::string, unsigned long long> const get_json_average_stats(unsigned long long elapsed_clock_ticks);
+  std::map<std::string, double> const get_json_average_stats(unsigned long long elapsed_clock_ticks);
 
 };
