+++ package/src/memmon.h
@@ -1,4 +1,4 @@
-// Copyright (C) CERN, 2018
+// Copyright (C) CERN, 2020
 //
 // Memory monitoring class
 //
@@ -20,7 +20,7 @@ class memmon final : public Imonitor {
   // Container for total stats
   std::map<std::string, unsigned long long> mem_stats;
   std::map<std::string, unsigned long long> mem_peak_stats;
-  std::map<std::string, unsigned long long> mem_average_stats;
+  std::map<std::string, double> mem_average_stats;
   std::map<std::string, unsigned long long> mem_total_stats;
 
   // Counter for number of iterations
@@ -35,7 +35,7 @@ class memmon final : public Imonitor {
   // These are the stat getter methods which retrieve current statistics
   std::map<std::string, unsigned long long> const get_text_stats();
   std::map<std::string, unsigned long long> const get_json_total_stats();
-  std::map<std::string, unsigned long long> const get_json_average_stats(unsigned long long elapsed_clock_ticks);
+  std::map<std::string, double> const get_json_average_stats(unsigned long long elapsed_clock_ticks);
 
 };
